moved todo type into a separate folder and export it as default
