--- src/modules/todo/components/todo/Todo.interface.ts
@@ -1,6 +1,0 @@
-interface TodoItem {
-  id: string
-  text: string
-}
-
-export type { TodoItem }
