+++ src/modules/todo/state/todo-slice.ts
@@ -1,6 +1,6 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
 import { RootState } from '../../../store'
-import type { TodoItem } from '../components/todo/Todo.interface'
+import type TodoItem from '../types/Todo.interface'
 
 interface TodoState {
   todos: Record<string, TodoItem>
