+++ src/modules/todo/types/Todo.interface.ts
@@ -1,0 +1,6 @@
+ interface TodoItem {
+  id: string
+  text: string
+}
+
+export default TodoItem
